# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Homeroom API
  version: 0.1.0
  description: |
    Contract shared by the Next.js web app and the Flutter app. Every route is scoped to the
    caller's school and role by Postgres row-level security (see backend/migrations); the
    "Roles" line on each operation says who is meant to call it, and the database enforces it.

    **Conventions**
    - Auth: `Authorization: Bearer <jwt>`. The API maps the token to `people.id` and sets `app.person_id`.
    - IDs are UUIDs. Dates are `YYYY-MM-DD`; timestamps are RFC 3339 UTC. Money is a decimal string in GHS minor-unit-safe form (`"1850.00"`), never a float.
    - Lists are cursor-paginated: `?limit=50&cursor=...` returns `{ items, next_cursor }`.
    - Errors use `application/problem+json`. A row the caller may not see is reported as **404**, not 403, so existence is not leaked. 403 is for a visible resource the caller may not change.
    - Writes from the offline apps send an `Idempotency-Key` header; replays return the original result.
servers:
  - url: https://api.homeroom.example/v1
  - url: http://localhost:3000/api/v1
security:
  - bearer: []
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [ Me ]
      summary: 'Who am I, and what can I switch between'
      description: |
        Roles: any. Guardians get `children` (drives the child switcher); teachers get `sections`.
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200':
          description: The caller
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
  /me/notification_prefs:
    get:
      operationId: getNotificationPrefs
      tags: [ Me ]
      summary: Channels the caller receives alerts on
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { description: 'Prefs', content: { application/json: { schema: { $ref: '#/components/schemas/NotificationPrefs' } } } }
    put:
      operationId: setNotificationPrefs
      tags: [ Me ]
      summary: Replace notification channel preferences
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/NotificationPrefs' } } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { description: 'Saved', content: { application/json: { schema: { $ref: '#/components/schemas/NotificationPrefs' } } } }
  /me/push_tokens:
    post:
      operationId: registerPushToken
      tags: [ Me ]
      summary: Register a device for push
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ token, platform ]
              properties:
                token: { type: string }
                platform: { type: string, enum: [ ios, android, web ] }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '204': { description: 'Registered (idempotent on token)' }
  /me/push_tokens/{token}:
    delete:
      operationId: unregisterPushToken
      tags: [ Me ]
      summary: Unregister a device (call on sign-out)
      parameters: [ { name: token, in: path, required: true, schema: { type: string } } ]
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '204': { description: 'Removed' }

  # ------------------------------------------------------------------ overview (screen-shaped)
...
  /students/{id}/homework:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      operationId: getStudentHomework
      tags: [ Homework ]
      summary: Homework for a child's sections
      description: 'Roles: guardian (own child), teacher, admin.'
      parameters: [ { name: due_from, in: query, schema: { type: string, format: date } } ]
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200':
          description: Homework
          content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Homework' } } }, required: [ items ] } } }
...
  /class_sections/{id}/homework:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      operationId: listHomework
      tags: [ Homework ]
      summary: Homework for a section
      description: 'Roles: teacher (taught), admin (view).'
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200':
          description: Homework
          content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Homework' } } }, required: [ items ] } } }
    post:
      operationId: postHomework
      tags: [ Homework ]
      summary: Post homework
      description: 'Roles: teacher (taught, term open). Notifies guardians of enrolled students.'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/HomeworkInput' } } } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '201': { description: 'Posted', content: { application/json: { schema: { $ref: '#/components/schemas/Homework' } } } }

  # ------------------------------------------------------------------ conferences
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [ Me ]
      summary: My in-app notifications, newest first
      description: 'Roles: any. Attendance alerts, fee receipts, homework, announcements, messages.'
      parameters:
        - { name: unread, in: query, schema: { type: boolean } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200': { description: 'Page', content: { application/json: { schema: { $ref: '#/components/schemas/NotificationPage' } } } }
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
  /notifications/read:
    post:
      operationId: markNotificationsRead
      tags: [ Me ]
      summary: Mark notifications read (given ids, or all when ids is omitted)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids: { type: array, items: { type: string, format: uuid } }
      responses:
        '204': { description: 'Marked' }
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
  /me/contact:
    put:
      operationId: updateMyContact
      tags: [ Me ]
      summary: Update my own phone and email
      description: 'Roles: any. Only the caller''s own contact row; nobody else can read it except admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string, nullable: true }
                email: { type: string, format: email, nullable: true }
      responses:
        '200': { description: 'Saved', content: { application/json: { schema: { $ref: '#/components/schemas/Contact' } } } }
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
...
  responses:
...
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Contact:
      type: object
      properties:
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
      required: [ phone, email ]
...
    Notification:
      type: object
      required: [ id, kind, title, body, created_at ]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, example: attendance.marked }
        title: { type: string }
        body: { type: string }
        data: { type: object, additionalProperties: true }
        read_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    NotificationPage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Notification' } }
        next_cursor: { type: string, nullable: true }
        unread_count: { type: integer }
      required: [ items, next_cursor, unread_count ]
    Problem:
      type: object
      required: [ status, title ]
      properties:
        status: { type: integer }
        title: { type: string }
        detail: { type: string }
        code: { type: string, example: term_closed }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
        request_id:
          type: string
          description: Matches the X-Request-Id response header and the server logs
...
    Role: { type: string, enum: [ admin, teacher, guardian, student ] }
    Me:
      type: object
      required: [ id, full_name, role, school, now ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        school:
          type: object
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            timezone: { type: string }
            term: { $ref: '#/components/schemas/Term' }
          required: [ id, name, timezone ]
        contact: { $ref: '#/components/schemas/Contact' }
        children: { type: array, items: { $ref: '#/components/schemas/StudentRef' }, description: 'Guardians only' }
        sections: { type: array, items: { $ref: '#/components/schemas/ClassSection' }, description: 'Teachers only' }
        now:
          type: string
          format: date-time
          description: The server's current time. Clients use it to correct a wrong device clock and to agree with the server about what 'today' is.
    NotificationPrefs:
      type: object
      required: [ push, email, sms ]
      properties: { push: { type: boolean }, email: { type: boolean }, sms: { type: boolean } }
    StudentRef:
      type: object
      required: [ id, full_name ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        class_section: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string } }, required: [ id, name ] }
...
    Person:
      type: object
      required: [ id, full_name, role ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        contact: { type: object, properties: { phone: { type: string }, email: { type: string } } }
        assignments:
          type: array
          description: 'Teachers, in the admin directory: what they teach, in the current term'
          items:
            type: object
            properties:
              class_section: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string } }, required: [ id, name ] }
              subject: { type: string }
            required: [ class_section, subject ]
...
    Term:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        starts_on: { type: string, format: date }
        ends_on: { type: string, format: date }
        closed: { type: boolean }
      required: [ id, name, starts_on, ends_on, closed ]
...
    ClassSection:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: 6A }
        grade_level: { type: string }
        term_id: { type: string, format: uuid }
        homeroom_teacher: { $ref: '#/components/schemas/Person' }
        subject: { type: string, description: 'Present on /me and /class_sections for a teacher: the subject they teach here' }
      required: [ id, name, grade_level, term_id ]
...
    Homework:
      type: object
      properties:
        id: { type: string, format: uuid }
        class_section_id: { type: string, format: uuid }
        subject: { type: string, nullable: true }
        title: { type: string }
        body: { type: string, nullable: true }
        due_date: { type: string, format: date }
        posted_by: { $ref: '#/components/schemas/Person' }
      required: [ id, class_section_id, subject, title, body, due_date ]
    HomeworkInput:
      type: object
      required: [ title, due_date ]
      properties:
        title: { type: string }
        body: { type: string }
        due_date: { type: string, format: date }
...